backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
pydantic
tqdm